+++ openbank-tpp-registry-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.9.1](https://github.com/JiRaska/open-bank-oss/compare/tpp-registry-service-v0.9.0...tpp-registry-service-v0.9.1) (2026-09-01)
+
+
+### Bug Fixes
+
+* **tpp-registry:** reject a null array element with 400 instead of 500 ([#8007](https://github.com/JiRaska/open-bank-oss/issues/8007)) ([b8dd50c](https://github.com/JiRaska/open-bank-oss/commit/b8dd50c8fe334b014c3e2818665f3628245bebd5))
+
 ## [0.9.0](https://github.com/JiRaska/open-bank-oss/compare/tpp-registry-service-v0.8.0...tpp-registry-service-v0.9.0) (2026-08-24)